Refactor profiles to clean up SkyModelPlot
